// src/app/models/academics/school-rep-role-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * SchoolRepRoleView
 */
export interface SchoolRepRoleView { 
    /**
     * Gets or sets  School RepRole Name
     */
    name: string;
    /**
     * Gets or Sets BaseViewModel
     */
    id?: string;
}